uint to address conversion fixed for compiler version 0.8.0, IERC20 used for transfers. FxERC20 safe transfer to be implemented. Not done.

+++ contracts/lib/RLPReader.sol
@@ -185,7 +185,7 @@ library RLPReader {
         // 1 byte for the length prefix
         require(item.len == 21);
 
-        return address(toUint(item));
+        return address(uint160(toUint(item)));
     }
 
     function toUint(RLPItem memory item) internal pure returns (uint) {
